LayoutTests/platform/mac/accessibility/div-containing-div-with-aria.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script>
var successfullyParsed = false;
</script>
<script src="../../../fast/js/resources/js-test-pre.js"></script>
</head>
<body id="body">

<div id="outer" aria-live="polite" aria-relevant="text" tabindex=0 aria-label="test">
   <div id="display" class="LCD">0</div>
</div>


<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that any element that has ARIA attributes is not ignored");

    if (window.accessibilityController) {

        document.getElementById("outer").focus();
        var outerDiv = accessibilityController.focusedElement;
        shouldBe("outerDiv.description", "'AXDescription: test'");
        shouldBe("outerDiv.childAtIndex(0).role", "'AXRole: AXGroup'");
        shouldBe("outerDiv.childAtIndex(0).childAtIndex(0).role", "'AXRole: AXStaticText'");
    }

    successfullyParsed = true;
</script>

<script src="../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
