safari-600.7-branch/LayoutTests/platform/mac/accessibility/div-containing-div-with-aria-expected.txt

0
This tests that any element that has ARIA attributes is not ignored

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS outerDiv.description is 'AXDescription: test'
PASS outerDiv.childAtIndex(0).role is 'AXRole: AXGroup'
PASS outerDiv.childAtIndex(0).childAtIndex(0).role is 'AXRole: AXStaticText'
PASS successfullyParsed is true

TEST COMPLETE
